refactor(adapters): unify createTestEngine entry points for 1.0
Gives every framework adapter the same entry-point shape so consumers find a
consistent `createTestEngine(subject, parts, option?)` across DOM, React, Vue and
Playwright, backed by a single shared option type in core.

## Key Changes

- Name: dom-core's `createDomTestEngine` is renamed to `createTestEngine` (the file
  moves to `createTestEngine.ts` for symmetry); `createDomTestEngine` stays as a
  deprecated alias.
- Arity: Playwright's `createTestEngine` gains an `option?` third parameter (a
  documented no-op — `rootElement` is not applicable to a real browser page) so all
  adapters share `createTestEngine(subject, parts, option?)`.
- Option type: add `ITestEngineOption extends IComponentDriverOption { rootElement?:
  Element }` to `@atomic-testing/core` (exported from the barrel). The previously
  byte-identical `IReactTestEngineOption` (×3), `IVueTestEngineOption` and
  `IDomTestEngineOption` become deprecated aliases of it — one extension point for
  future universal bootstrap options.

Verified: full typecheck + recursive build green; html (react-19) and vue-3 dom
suites pass.

Closes #971
