+++ packages/dom-core/src/index.ts
@@ -1,4 +1,4 @@
 export { DOMInteractor } from './DOMInteractor';
-export * from './createDomTestEngine';
+export * from './createTestEngine';
 export * from './fakeEvents';
 export * from './types';
